// src/interfaces/AnimeRecomendations.tsx
// Generated by https://quicktype.io

export interface AnimeRecomendations {
    pagination: Pagination;
    data:       Datum[];
}

export interface Datum {
    mal_id:  string;
    entry:   Entry[];
    content: string;
    date:    string;
    user:    User;
}

export interface Entry {
    mal_id: number;
    url:    string;
    images: { [key: string]: Image };
    title:  string;
}

export interface Image {
    image_url:       string;
    small_image_url: string;
    large_image_url: string;
}

export interface User {
    url:      string;
    username: string;
}

export interface Pagination {
    last_visible_page: number;
    has_next_page:     boolean;
}